# src/main/resources/contracts/boards-preview.openapi.yaml
openapi: 3.0.3
info:
  title: Weave Boards/Tasks Preview Contract
  version: preview
  description: >-
    Preview-only provider-neutral schema components for future Boards/Tasks work.
    Hidden preview-only provider-neutral schema and route draft for future Boards/Tasks work.
    The runtime routes are disabled by default, require the normal authenticated Weave
    workspace boundary, use only a local/in-memory preview adapter, and do not publish
    Boards/Tasks as a Release 1 product API.
paths:
  /api/boards/preview:
    get:
      summary: Read hidden Boards/Tasks preview snapshot
      responses:
        '200':
          description: Provider-neutral hidden preview snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardsPreviewResponse'
        '401': { description: Missing or invalid bearer token. }
        '403': { description: Missing weave:workspace scope. }
        '503': { description: Preview runtime is disabled or unavailable. }
  /api/boards/{boardId}/tasks:
    post:
      summary: Create a hidden preview task
      parameters:
        - name: boardId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardsCreateTaskRequest'
      responses:
        '200':
          description: Created provider-neutral task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'
  /api/boards/tasks/{taskId}/move:
    post:
      summary: Move a hidden preview task without drag-and-drop
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardsMoveTaskRequest'
      responses:
        '200':
          description: Moved provider-neutral task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'
  /api/boards/tasks/{taskId}/complete:
    post:
      summary: Complete a hidden preview task without drag-and-drop
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Completed provider-neutral task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'

components:
  schemas:

    BoardsPreviewResponse:
      type: object
      required: [preview, releaseStatus, source, capabilities, projects, boards, tasks]
      properties:
        preview: { type: boolean }
        releaseStatus:
          type: string
          enum: [post-release-hidden-preview]
        source:
          type: string
          enum: [local-preview-backend-facade]
        capabilities:
          $ref: '#/components/schemas/BoardProviderCapabilities'
        projects:
          type: array
          items: { $ref: '#/components/schemas/WeaveProject' }
        boards:
          type: array
          items: { $ref: '#/components/schemas/Board' }
        tasks:
          type: array
          items: { $ref: '#/components/schemas/TaskItem' }
    BoardProviderCapabilities:
      type: object
      required: [provider, enabled, supported, unsupported, supportSafeSummary]
      properties:
        provider:
          type: string
          enum: [vikunja, openproject, nextcloud-deck, in-memory, unknown]
        enabled: { type: boolean }
        supported:
          type: array
          items:
            type: string
            enum: [comments, attachments, non_destructive_archive, webhook_events, incremental_sync, checklists, custom_fields, accessible_non_drag_moves]
        unsupported:
          type: array
          items:
            type: string
            enum: [comments, attachments, non_destructive_archive, webhook_events, incremental_sync, checklists, custom_fields, accessible_non_drag_moves]
        supportSafeSummary: { type: string }
    BoardsCreateTaskRequest:
      type: object
      required: [columnId, title]
      properties:
        columnId: { type: string, maxLength: 128 }
        title: { type: string, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        assigneeRefs:
          type: array
          items: { type: string, maxLength: 128 }
        labelRefs:
          type: array
          items: { type: string, maxLength: 128 }
        dueAt: { type: string, format: date-time }
    BoardsMoveTaskRequest:
      type: object
      required: [targetColumnId, targetPosition]
      properties:
        targetColumnId: { type: string, maxLength: 128 }
        targetPosition: { type: integer, minimum: 0 }

    ProviderRef:
      type: object
      required: [provider, externalId]
      properties:
        provider:
          type: string
          enum: [vikunja, openproject, nextcloud-deck, in-memory, unknown]
        externalId:
          type: string
        externalUrl:
          type: string
          format: uri
        version:
          type: string
        etag:
          type: string
        lastSyncedAt:
          type: string
          format: date-time
    WeaveProject:
      type: object
      required: [id, name, visibility, memberRefs, providerRefs]
      properties:
        id: { type: string }
        name: { type: string }
        visibility:
          type: string
          enum: [private, workspace, public]
        memberRefs:
          type: array
          items: { type: string }
        providerRefs:
          type: array
          items: { $ref: '#/components/schemas/ProviderRef' }
    Board:
      type: object
      required: [id, projectId, name, columns, archived, providerRefs]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        description: { type: string }
        columns:
          type: array
          items: { $ref: '#/components/schemas/BoardColumn' }
        archived: { type: boolean }
        providerRefs:
          type: array
          items: { $ref: '#/components/schemas/ProviderRef' }
    BoardColumn:
      type: object
      required: [id, boardId, name, position, semanticStatus, providerRefs]
      properties:
        id: { type: string }
        boardId: { type: string }
        name: { type: string }
        position: { type: integer, minimum: 0 }
        semanticStatus:
          type: string
          enum: [not_started, in_progress, blocked, done, archived]
        wipLimit:
          type: integer
          minimum: 1
        providerRefs:
          type: array
          items: { $ref: '#/components/schemas/ProviderRef' }
    TaskItem:
      type: object
      required: [id, boardId, columnId, title, status, position, assigneeRefs, labelRefs, updatedAt, providerRefs]
      properties:
        id: { type: string }
        boardId: { type: string }
        columnId: { type: string }
        title: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [open, blocked, completed, archived]
        position: { type: integer, minimum: 0 }
        assigneeRefs:
          type: array
          items: { type: string }
        labelRefs:
          type: array
          items: { type: string }
        priority:
          type: string
          enum: [low, normal, high, urgent]
        startAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        providerRefs:
          type: array
          items: { $ref: '#/components/schemas/ProviderRef' }
    TaskBoardEvent:
      type: object
      required: [idempotencyKey, type, actorRef, occurredAt, workspaceId, redactionLevel, payload]
      properties:
        idempotencyKey: { type: string }
        type:
          type: string
          enum:
            - task.created
            - task.updated
            - task.completed
            - task.archived
            - task.moved
            - assignment.changed
            - label.changed
            - priority.changed
            - due_date.changed
            - comment.added
            - attachment.changed
            - sync.conflict_detected
        actorRef: { type: string }
        occurredAt: { type: string, format: date-time }
        workspaceId: { type: string }
        projectId: { type: string }
        boardId: { type: string }
        taskId: { type: string }
        providerRef: { $ref: '#/components/schemas/ProviderRef' }
        redactionLevel:
          type: string
          enum: [public_metadata, workspace_internal, support_safe, private_content]
        payload:
          type: object
          additionalProperties: true